+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Added
   * AZ-694: Add log analytics workspace parameter + site config parameter
   * AZ-694: Add additional function-app options (certificat requierement + built-in login)
 
+Breaking
+  * AZ-588: Storage account network rules
+
 # v5.1.0 - 2022-02-11
 
 Added